feat!: generic Ring[T] core + RingBuffer byte wrapper, Go 1.23 modernization
BREAKING CHANGES:
- Introduce generic core type Ring[T any]
- RingBuffer becomes a byte-specific wrapper around *Ring[byte]
- Byte-specific methods (WriteString, ReadFrom, WriteTo, Copy, Pipe)
  remain on *RingBuffer for full io compatibility
- Requires Go 1.23+

FEATURES:
- Add All() returning iter.Seq[T] (Go 1.23 iterators)
- Add NewRing[T], NewRingFromSlice[T] constructors
- RingBuffer fully preserves original API including io.ByteReader,
  io.ByteWriter, io.ReaderFrom, io.WriterTo, io.StringWriter

FIXES:
- Fix goroutine leak in WithCancel using done channel + sync.Once
- Modernize unsafe string-to-bytes conversion in WriteString
- Use builtin min/max and clear() for cleaner code

INFRA:
- Update module path to github.com/argcv/ringbuffer
- Update CI to Go 1.23 and latest action versions
- Remove legacy .travis.yml
- Rewrite README with new API docs
