+++ go.mod
@@ -1,3 +1,3 @@
-module github.com/smallnest/ringbuffer
+module github.com/argcv/ringbuffer
 
-go 1.19
+go 1.23.0
